Refactor DuplicatedEventTypeNameException to be uncheked exception
