+++ src/main/java/org/zalando/nakadi/exceptions/runtime/DuplicatedEventTypeNameException.java
@@ -1,0 +1,12 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class DuplicatedEventTypeNameException extends MyNakadiRuntimeException1 {
+
+    public DuplicatedEventTypeNameException(final String msg, final Exception cause) {
+        super(msg, cause);
+    }
+
+    public DuplicatedEventTypeNameException(final String msg) {
+        super(msg);
+    }
+}
